+++ lib/twilio/device.ts
@@ -5,6 +5,7 @@
  * @publicapi
  */
 import { EventEmitter } from 'events';
+import { levels as LogLevels, LogLevelDesc } from 'loglevel';
 import AudioHelper from './audiohelper';
 import Connection from './connection';
 import DialtonePlayer from './dialtonePlayer';
@@ -12,6 +13,8 @@ import {
   AuthorizationErrors,
   ClientErrors,
   GeneralErrors,
+  getErrorByCode,
+  hasErrorByCode,
   InvalidArgumentError,
   InvalidStateError,
   NotSupportedError,
@@ -59,9 +62,6 @@ const RINGTONE_PLAY_TIMEOUT = 2000;
 declare const RTCRtpTransceiver: any;
 declare const webkitAudioContext: typeof AudioContext;
 
-let hasBeenWarnedHandlers: boolean = false;
-let hasBeenWarnedSounds: boolean = false;
-
 /**
  * Options that may be passed to the {@link Device} constructor for internal testing.
  * @private
@@ -91,12 +91,6 @@ export interface IExtendedDeviceOptions extends Device.Options {
    * File input stream to use instead of reading from mic
    */
   fileInputStream?: MediaStream;
-
-  /**
-   * A list of specific ICE servers to use. Overridden by {@link Device.Options.rtcConfiguration}.
-   * @deprecated
-   */
-  iceServers?: Object[];
 
   /**
    * Ignore browser support, disabling the exception that is thrown when neither WebRTC nor
@@ -363,18 +357,15 @@ class Device extends EventEmitter {
     closeProtection: false,
     codecPreferences: [Connection.Codec.PCMU, Connection.Codec.Opus],
     connectionFactory: Connection,
-    debug: false,
     dscp: true,
-    enableIceRestart: false,
     eventgw: 'eventgw.twilio.com',
     forceAggressiveIceNomination: false,
-    iceServers: [],
+    logLevel: LogLevels.ERROR,
     noRegister: false,
     pStreamFactory: PStream,
     preflight: false,
     soundFactory: Sound,
     sounds: { },
-    warnings: true,
   };
 
   /**
@@ -440,21 +431,7 @@ class Device extends EventEmitter {
    * Return the active {@link Connection}. Null or undefined for backward compatibility.
    */
   activeConnection(): Connection | null | undefined {
-    if (!this.isInitialized) {
-      return null;
-    }
-    // @rrowland This should only return activeConnection, but customers have built around this
-    // broken behavior and in order to not break their apps we are including this until
-    // the next big release.
-    return this._activeConnection || this.connections[0];
-  }
-
-  /**
-   * @deprecated Set a handler for the cancel event.
-   * @param handler
-   */
-  cancel(handler: (connection: Connection) => any): this {
-    return this._addHandler(Device.EventName.Cancel, handler);
+    return this.isInitialized ? this._activeConnection : null;
   }
 
   /**
@@ -513,15 +490,6 @@ class Device extends EventEmitter {
   }
 
   /**
-   * Set a handler for the disconnect event.
-   * @deprecated Use {@link Device.on}.
-   * @param handler
-   */
-  disconnect(handler: (connection: Connection) => any): this {
-    return this._addHandler(Device.EventName.Disconnect, handler);
-  }
-
-  /**
    * Disconnect all {@link Connection}s.
    */
   disconnectAll(): void {
@@ -535,55 +503,6 @@ class Device extends EventEmitter {
    */
   get edge(): string | null {
     return this._edge;
-  }
-
-  /**
-   * Set a handler for the error event.
-   * @deprecated Use {@link Device.on}.
-   * @param handler
-   */
-  error(handler: (error: Connection) => any): this {
-    return this._addHandler(Device.EventName.Error, handler);
-  }
-
-  /**
-   * Set a handler for the incoming event.
-   * @deprecated Use {@link Device.on}.
-   * @param handler
-   */
-  incoming(handler: (connection: Connection) => any): this {
-    return this._addHandler(Device.EventName.Incoming, handler);
-  }
-
-  /**
-   * Set a handler for the offline event.
-   * @deprecated Use {@link Device.on}.
-   * @param handler
-   */
-  offline(handler: (device: Device) => any): this {
-    return this._addHandler(Device.EventName.Offline, handler);
-  }
-
-  /**
-   * Set a handler for the ready event.
-   * @deprecated Use {@link Device.on}.
-   * @param handler
-   */
-  ready(handler: (device: Device) => any): this {
-    return this._addHandler(Device.EventName.Ready, handler);
-  }
-
-  /**
-   * Get the {@link Region} string the {@link Device} is currently connected to, or 'offline'
-   * if not connected.
-   */
-  region(): string {
-    this._log.warn(
-      '`Device.region` is deprecated and will be removed in the next major ' +
-      'release. Please use `Device.edge` instead.',
-    );
-    this._throwUnlessSetup('region');
-    return typeof this._region === 'string' ? this._region : 'offline';
   }
 
   /**
@@ -642,18 +561,16 @@ class Device extends EventEmitter {
     Object.assign(this.options, options);
 
     this._log.setDefaultLevel(
-      this.options.debug
-        ? Log.levels.DEBUG
-        : this.options.warnings
-          ? Log.levels.WARN
-          : Log.levels.SILENT,
+      typeof this.options.logLevel === 'number'
+        ? this.options.logLevel
+        : LogLevels.ERROR,
     );
 
     this._chunderURIs = this.options.chunderw
       ? [`wss://${this.options.chunderw}/signal`]
       : getChunderURIs(
           this.options.edge,
-          this.options.region,
+          undefined,
           this._log.warn.bind(this._log),
         ).map((uri: string) => `wss://${uri}/signal`);
 
@@ -667,7 +584,7 @@ class Device extends EventEmitter {
 
     Device._getOrCreateAudioContext();
 
-    if (Device._audioContext && options.fakeLocalDTMF) {
+    if (Device._audioContext) {
       if (!Device._dialtonePlayer) {
         Device._dialtonePlayer = new DialtonePlayer(Device._audioContext);
       }
@@ -684,24 +601,9 @@ class Device extends EventEmitter {
 
     this.isInitialized = true;
 
-    const getOrSetSound = (key: Device.ToggleableSound, value?: boolean) => {
-      if (!hasBeenWarnedSounds) {
-        this._log.warn('Device.sounds is deprecated and will be removed in the next breaking ' +
-          'release. Please use the new functionality available on Device.audio.');
-        hasBeenWarnedSounds = true;
-      }
-
-      if (typeof value !== 'undefined') {
-        this._enabledSounds[key] = value;
-      }
-
-      return this._enabledSounds[key];
-    };
-
-    [Device.SoundName.Disconnect, Device.SoundName.Incoming, Device.SoundName.Outgoing]
-        .forEach((eventName: Device.SoundName) => {
-      this.sounds[eventName] = getOrSetSound.bind(null, eventName);
-    });
+    if (this.options.dscp) {
+      (this.options.rtcConstraints as any).optional = [{ googDscp: true }];
+    }
 
     const defaultSounds: Record<string, ISoundDefinition> = {
       disconnect: { filename: 'disconnect', maxDuration: 3000 },
@@ -789,15 +691,6 @@ class Device extends EventEmitter {
       }
     }
 
-    // (rrowland) This maintains backward compatibility, but we should look at
-    // removing this next breaking change. Any error should be caught by the
-    // customer, and anything that's not a fatal error should not be emitted
-    // via error event.
-    this.on(Device.EventName.Error, () => {
-      if (this.listenerCount('error') > 1) { return; }
-      this._log.info('Uncaught error event suppressed.');
-    });
-
     return this;
   }
 
@@ -836,23 +729,6 @@ class Device extends EventEmitter {
     this._throwUnlessSetup('updateToken');
     this.token = token;
     this.register(token);
-  }
-
-  /**
-   * Add a handler for an EventEmitter and emit a deprecation warning on first call.
-   * @param eventName - Name of the event
-   * @param handler - A handler to call when the event is emitted
-   */
-  private _addHandler(eventName: Device.EventName, handler: (...args: any[]) => any): this {
-    if (!hasBeenWarnedHandlers) {
-      this._log.warn(`Device callback handlers (connect, error, offline, incoming, cancel, ready, disconnect) \
-        have been deprecated and will be removed in the next breaking release. Instead, the EventEmitter \
-        interface can be used to set event listeners. Example: device.on('${eventName}', handler)`);
-      hasBeenWarnedHandlers = true;
-    }
-
-    this.addListener(eventName, handler);
-    return this;
   }
 
   /**
@@ -895,7 +771,7 @@ class Device extends EventEmitter {
       aggressive_nomination: this.options.forceAggressiveIceNomination,
       browser_extension: this._isBrowserExtension,
       dscp: !!this.options.dscp,
-      ice_restart_enabled: this.options.enableIceRestart,
+      ice_restart_enabled: true,
       platform: rtc.getMediaEngine(),
       sdk_version: C.RELEASE_VERSION,
     };
@@ -913,7 +789,6 @@ class Device extends EventEmitter {
     }
 
     setIfDefined('gateway', this.stream && this.stream.gateway);
-    setIfDefined('selected_region', this.options.region);
     setIfDefined('region', this.stream && this.stream.region);
 
     return payload;
@@ -974,13 +849,12 @@ class Device extends EventEmitter {
       codecPreferences: this.options.codecPreferences,
       dialtonePlayer: Device._dialtonePlayer,
       dscp: this.options.dscp,
-      enableIceRestart: this.options.enableIceRestart,
-      enableRingingState: this.options.enableRingingState,
       forceAggressiveIceNomination: this.options.forceAggressiveIceNomination,
       getInputStream: (): MediaStream | null => this.options.fileInputStream || this._connectionInputStream,
       getSinkIds: (): string[] => this._connectionSinkIds,
       maxAverageBitrate: this.options.maxAverageBitrate,
       preflight: this.options.preflight,
+      rtcConstraints: this.options.rtcConstraints,
       shouldPlayDisconnect: () => this._enabledSounds.disconnect,
       twimlParams,
     }, options);
@@ -1090,28 +964,39 @@ class Device extends EventEmitter {
    * Called when an 'error' event is received from the signaling stream.
    */
   private _onSignalingError = (payload: Record<string, any>) => {
-    if (!payload.error) { return; }
+    if (typeof payload !== 'object') { return; }
 
-    const error = { ...payload.error };
-    const sid = payload.callsid;
-    if (sid) {
-      error.connection = this._findConnection(sid);
+    const { error, callsid } = payload;
+
+    if (typeof error !== 'object') { return; }
+
+    const connection: Connection | undefined =
+      (typeof callsid === 'string' && this._findConnection(callsid)) || undefined;
+
+    const { code } = error;
+    let { twilioError } = error;
+
+    if (!twilioError && typeof code === 'number') {
+      if (code === 31201) {
+        twilioError = new AuthorizationErrors.AuthenticationFailed();
+      } else if (code === 31204) {
+        twilioError = new AuthorizationErrors.AccessTokenInvalid();
+      } else if (code === 31205) {
+        // Stop trying to register presence after token expires
+        this._stopRegistrationTimer();
+        twilioError = new AuthorizationErrors.AccessTokenExpired();
+      } else if (hasErrorByCode(code)) {
+        twilioError = new (getErrorByCode(code))();
+      }
     }
 
-    if (error.code === 31201) {
-      error.twilioError = new AuthorizationErrors.AuthenticationFailed();
-    } else if (error.code === 31204) {
-      error.twilioError = new AuthorizationErrors.AccessTokenInvalid();
-    } else if (error.code === 31205) {
-      // Stop trying to register presence after token expires
-      this._stopRegistrationTimer();
-      error.twilioError = new AuthorizationErrors.AccessTokenExpired();
-    } else if (!error.twilioError) {
-      error.twilioError = new GeneralErrors.UnknownError();
+    if (!twilioError) {
+      this._log.error('Unknown signaling error: ', error);
+      twilioError = new GeneralErrors.UnknownError();
     }
 
-    this._log.info('Received error: ', error);
-    this.emit('error', error);
+    this._log.info('Received error: ', twilioError);
+    this.emit('error', twilioError, connection);
   }
 
   /**
@@ -1125,7 +1010,7 @@ class Device extends EventEmitter {
     }
 
     if (!payload.callsid || !payload.sdp) {
-      this.emit('error', { message: 'Malformed invite from gateway', twilioError: new ClientErrors.BadRequest() });
+      this.emit('error', new ClientErrors.BadRequest('Malformed invite from gateway'));
       return;
     }
 
@@ -1401,7 +1286,7 @@ namespace Device {
    * @example `device.on('error', connection => { })`
    * @event
    */
-  declare function errorEvent(error: Connection): void;
+  declare function errorEvent(error: TwilioError, connection?: Connection): void;
 
   /**
    * Emitted when an incoming {@link Connection} is received.
@@ -1555,11 +1440,6 @@ namespace Device {
     codecPreferences?: Connection.Codec[];
 
     /**
-     * Whether to enable debug logging.
-     */
-    debug?: boolean;
-
-    /**
      * Whether AudioContext sounds should be disabled. Useful for trouble shooting sound issues
      * that may be caused by AudioContext-specific sounds. If set to true, will fall back to
      * HTMLAudioElement sounds.
@@ -1575,35 +1455,20 @@ namespace Device {
      * The edge value corresponds to the geographic location that the client
      * will use to connect to Twilio infrastructure. The default value is
      * "roaming" which automatically selects an edge based on the latency of the
-     * client relative to available edges. You may not specify both `edge` and
-     * `region` in the Device options. Specifying both `edge` and `region` will
-     * result in an `InvalidArgumentException`.
+     * client relative to available edges.
      */
     edge?: string[] | string;
-
-    /**
-     * Whether to automatically restart ICE when media connection fails
-     */
-    enableIceRestart?: boolean;
-
-    /**
-     * Whether the ringing state should be enabled on {@link Connection} objects. This is required
-     * to enable answerOnBridge functionality.
-     */
-    enableRingingState?: boolean;
-
-    /**
-     * Whether or not to override the local DTMF sounds with fake dialtones. This won't affect
-     * the DTMF tone sent over the connection, but will prevent double-send issues caused by
-     * using real DTMF tones for user interface. In 2.0, this will be enabled by default.
-     */
-    fakeLocalDTMF?: boolean;
 
     /**
      * Experimental feature.
      * Whether to use ICE Aggressive nomination.
      */
     forceAggressiveIceNomination?: boolean;
+
+    /**
+     * Log level.
+     */
+    logLevel?: LogLevelDesc;
 
     /**
      * The maximum average audio bitrate to use, in bits per second (bps) based on
@@ -1616,48 +1481,9 @@ namespace Device {
     maxAverageBitrate?: number;
 
     /**
-     * The region code of the region to connect to.
-     *
-     * @deprecated
-     *
-     * CLIENT-7519 This parameter is deprecated in favor of the `edge`
-     * parameter. You may not specify both `edge` and `region` in the Device
-     * options.
-     *
-     * This parameter will be removed in the next major version release.
-     *
-     * The following table lists the new edge names to region name mappings.
-     * Instead of passing the `region` value in `options.region`, please pass the
-     * following `edge` value in `options.edge`.
-     *
-     * | Region Value | Edge Value   |
-     * |:-------------|:-------------|
-     * | au1          | sydney       |
-     * | br1          | sao-paulo    |
-     * | ie1          | dublin       |
-     * | de1          | frankfurt    |
-     * | jp1          | tokyo        |
-     * | sg1          | singapore    |
-     * | us1          | ashburn      |
-     * | us2          | umatilla     |
-     * | gll          | roaming      |
-     * | us1-ix       | ashburn-ix   |
-     * | us2-ix       | san-jose-ix  |
-     * | ie1-ix       | london-ix    |
-     * | de1-ix       | frankfurt-ix |
-     * | sg1-ix       | singapore-ix |
-     */
-    region?: string;
-
-    /**
      * A mapping of custom sound URLs by sound name.
      */
     sounds?: Partial<Record<Device.SoundName, string>>;
-
-    /**
-     * Whether to enable warn logging.
-     */
-    warnings?: boolean;
   }
 }