+++ tests/integration/stirshaken.ts
@@ -2,7 +2,7 @@ import Connection from '../../lib/twilio/connection';
 import Device from '../../lib/twilio/device';
 import { generateAccessToken } from '../lib/token';
 import * as assert from 'assert';
-import { EventEmitter } from 'events';    
+import { EventEmitter } from 'events';
 import * as env from '../env';
 
 describe('SHAKEN/STIR', function() {
@@ -24,14 +24,24 @@ describe('SHAKEN/STIR', function() {
     device1 = new Device();
     device2 = new Device();
 
-    options = {
-      warnings: false,
-    };
+    options = {};
 
     return Promise.all([
       expectEvent('ready', device1.setup(token1, options)),
       expectEvent('ready', device2.setup(token2, options)),
     ]);
+  });
+
+  after(() => {
+    if (device1) {
+      device1.disconnectAll();
+      device1.destroy();
+    }
+
+    if (device2) {
+      device2.disconnectAll();
+      device2.destroy();
+    }
   });
 
   describe('device 1 calls device 2', () => {
@@ -63,8 +73,8 @@ describe('SHAKEN/STIR', function() {
         let connection2: Connection;
 
         beforeEach(() => {
-          const conn1: Connection | undefined | null = device1.activeConnection();
-          const conn2: Connection | undefined | null = device2.activeConnection();
+          const conn1: Connection | undefined | null = device1.activeConnection() || device1.connections[0];
+          const conn2: Connection | undefined | null = device2.activeConnection() || device2.connections[0];
 
           if (!conn1 || !conn2) {
             throw new Error(`Connections weren't both open at beforeEach`);